fix: remove other builder models
